backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.136.0",
    "pydantic>=2.13.2",
    "python-dotenv>=1.2.2",
    "redis>=5.0.0",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn[standard]>=0.44.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
